move react components to their own folder
