+++ templates/datasource/src/components/ConfigEditor.tsx
@@ -1,7 +1,7 @@
 import React, { ChangeEvent, PureComponent } from 'react';
 import { LegacyForms } from '@grafana/ui';
 import { DataSourcePluginOptionsEditorProps } from '@grafana/data';
-import { MyDataSourceOptions, MySecureJsonData } from './types';
+import { MyDataSourceOptions, MySecureJsonData } from '../types';
 
 const { SecretFormField, FormField } = LegacyForms;
