+++ templates/datasource/src/components/QueryEditor.tsx
@@ -3,8 +3,8 @@ import defaults from 'lodash/defaults';
 import React, { ChangeEvent, PureComponent } from 'react';
 import { LegacyForms } from '@grafana/ui';
 import { QueryEditorProps } from '@grafana/data';
-import { DataSource } from './datasource';
-import { defaultQuery, MyDataSourceOptions, MyQuery } from './types';
+import { DataSource } from '../datasource';
+import { defaultQuery, MyDataSourceOptions, MyQuery } from '../types';
 
 const { FormField } = LegacyForms;
