+++ templates/panel/src/components/SimplePanel.tsx
